refactor(client): use feature-based organization for components
