+++ apps/client/src/components/layout/app-header.tsx
@@ -1,6 +1,6 @@
 import { NavLink } from "react-router";
 
-import { APP_TITLE, NAV_LINKS } from "#/components/config/ui-constants";
+import { APP_TITLE, NAV_LINKS } from "#/config/ui-constants";
 import { cn } from "#/lib/utils";
 
 /** Top navigation bar with app title and page links. */
