+++ apps/client/src/features/game/components/viewport.tsx
