+++ apps/client/src/features/lobby/components/player-info.tsx
