+++ apps/client/src/routes/user.tsx
@@ -2,8 +2,8 @@ import type { FormEvent } from "react";
 import { useState } from "react";
 import { useNavigate } from "react-router";
 
-import { AuthForm } from "#/components/auth/auth-form";
 import { PageContainer } from "#/components/layout/page-container";
+import { AuthForm } from "#/features/auth/components/auth-form";
 import { useUserAuthStore } from "#/features/auth/store/user-auth-store";
 import { useMatchConnectionStore } from "#/features/match/store/match-connection-store";
