+++ apps/client/tests/components/auth/auth-form.test.tsx
@@ -4,7 +4,7 @@ import { cleanup, render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { afterEach, describe, expect, it, vi } from "vitest";
 
-import { AuthForm } from "#/components/auth/auth-form";
+import { AuthForm } from "#/features/auth/components/auth-form";
 
 const defaultProps = {
   displayName: "Commander",
